SPT-1969: Renamed UnrecoverableErrorLambda
Renamed UnrecoverableErrorLambda to GetUnrecoverableErrorHandler, which
now follows our preferred naming strategy.
