+++ infrastructure/identity-reuse-service/template.yaml
@@ -1377,7 +1377,43 @@ Resources:
         - Key: Name
           Value: !Sub "/aws/lambda/${AWS::StackName}-PostConfirmDetailsHandler"
 
-  UnrecoverableErrorRole:
+  GetUnrecoverableErrorHandler:
+    Type: AWS::Serverless::Function
+    Condition: IsNotProduction
+    Properties:
+      FunctionName: !Sub "${AWS::StackName}-GetUnrecoverableErrorHandler"
+      Handler: get-unrecoverable-error-handler.lambdaHandler
+      Role: !GetAtt GetUnrecoverableErrorHandlerRole.Arn
+      Environment:
+        Variables:
+          NODE_OPTIONS: "--enable-source-maps"
+      Layers:
+        - !Ref GovUKFrontEndLayer
+        - !If
+          - DynaTraceEnabled
+          - !Ref DynatraceNodeLayerArn
+          - !Ref AWS::NoValue
+    Metadata:
+      BuildMethod: esbuild
+      BuildProperties:
+        UseNpmCi: True
+        EntryPoints:
+          - src/handlers/get-unrecoverable-error-handler/get-unrecoverable-error-handler.ts
+        Sourcemap: true
+        SourcesContent: false
+        Target: "es2024"
+        Format: esm
+        Platform: node
+        OutExtension:
+          - .js=.mjs
+        Banner:
+          - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
+        Loader:
+          - .njk=file
+        External:
+          - govuk-frontend
+
+  GetUnrecoverableErrorHandlerRole:
     # checkov:skip=CKV_AWS_111: Ensure IAM policies does not allow write access without constraints
     Type: AWS::IAM::Role
     Condition: IsNotProduction
@@ -1398,48 +1434,12 @@ Resources:
         - arn:aws:iam::aws:policy/AmazonS3ReadOnlyAccess
         - arn:aws:iam::aws:policy/service-role/AWSLambdaVPCAccessExecutionRole
 
-  UnrecoverableErrorLambda:
-    Type: AWS::Serverless::Function
-    Condition: IsNotProduction
-    Properties:
-      FunctionName: !Sub "${AWS::StackName}-UnrecoverableErrorLambda"
-      Handler: unrecoverable-error-handler.lambdaHandler
-      Role: !GetAtt UnrecoverableErrorRole.Arn
-      Environment:
-        Variables:
-          NODE_OPTIONS: "--enable-source-maps"
-      Layers:
-        - !Ref GovUKFrontEndLayer
-        - !If
-          - DynaTraceEnabled
-          - !Ref DynatraceNodeLayerArn
-          - !Ref AWS::NoValue
-    Metadata:
-      BuildMethod: esbuild
-      BuildProperties:
-        UseNpmCi: True
-        EntryPoints:
-          - src/handlers/unrecoverable-error/unrecoverable-error-handler.ts
-        Sourcemap: true
-        SourcesContent: false
-        Target: "es2024"
-        Format: esm
-        Platform: node
-        OutExtension:
-          - .js=.mjs
-        Banner:
-          - js=import { createRequire } from 'module'; const require = createRequire(import.meta.url);
-        Loader:
-          - .njk=file
-        External:
-          - govuk-frontend
-
-  UnrecoverableErrorLambdaRestApiRole:
+  GetUnrecoverableErrorHandlerRestApiRole:
     Type: AWS::IAM::Role
     Condition: IsNotProduction
     Properties:
-      RoleName: !Sub "${AWS::StackName}-UnrecoverableErrorLambdaRestApiRole"
-      Description: !Sub "Role for the UnrecoverableErrorLambda"
+      RoleName: !Sub "${AWS::StackName}-GetUnrecoverableErrorHandlerRestApiRole"
+      Description: !Sub "Role for the GetUnrecoverableErrorHandler"
       PermissionsBoundary: !If [ UsePermissionsBoundary, !Ref PermissionsBoundary, !Ref AWS::NoValue ]
       AssumeRolePolicyDocument:
         Version: 2012-10-17
@@ -1461,11 +1461,11 @@ Resources:
         - Key: Source
           Value: !Ref SourceTagValue
 
-  UnrecoverableErrorLambdaRestApiPolicy:
+  GetUnrecoverableErrorHandlerRestApiPolicy:
     Type: AWS::IAM::Policy
     Condition: IsNotProduction
     Properties:
-      PolicyName: !Sub "${AWS::StackName}-UnrecoverableErrorLambdaRestApiPolicy"
+      PolicyName: !Sub "${AWS::StackName}-GetUnrecoverableErrorHandlerRestApiPolicy"
       PolicyDocument:
         Version: 2012-10-17
         Statement:
@@ -1473,15 +1473,15 @@ Resources:
             Action:
               - "lambda:InvokeFunction"
             Resource:
-              - !Sub "arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${AWS::StackName}-UnrecoverableErrorLambda:*"
+              - !Sub "arn:aws:lambda:${AWS::Region}:${AWS::AccountId}:function:${AWS::StackName}-GetUnrecoverableErrorHandler:*"
       Roles:
-        - !Ref UnrecoverableErrorLambdaRestApiRole
+        - !Ref GetUnrecoverableErrorHandlerRestApiRole
 
-  UnrecoverableErrorFunctionLogGroup:
+  GetUnrecoverableErrorHandlerLogGroup:
     Type: AWS::Logs::LogGroup
     Condition: IsNotProduction
     Properties:
-      LogGroupName: !Sub "/aws/lambda/${UnrecoverableErrorLambda}"
+      LogGroupName: !Sub "/aws/lambda/${AWS::StackName}-GetUnrecoverableErrorHandler"
       RetentionInDays: 30
       KmsKeyId: !GetAtt CloudWatchEncryptionKey.Arn
       Tags:
@@ -1496,7 +1496,7 @@ Resources:
         - Key: Source
           Value: !Ref SourceTagValue
         - Key: Name
-          Value: !Sub "/aws/lambda/${UnrecoverableErrorLambda}"
+          Value: !Sub "/aws/lambda/${AWS::StackName}-GetUnrecoverableErrorHandler"
 
   ################################
   # PRIVATE API API GATEWAY       #
