+++ src/handlers/get-unrecoverable-error-handler/get-unrecoverable-error-handler.ts
